rag-agent-guardrailed/requirements.txt
fastapi
uvicorn
langgraph>=0.2
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
langgraph-openai-serve
python-dotenv
aiohttp
vijil-dome